+++ R/utils.R
@@ -222,10 +222,14 @@ calc_sd_poly <- function(y, degree = 1) {
   stats::sd(model$residual)
 }
 
-# Compute the vector cross product between x and y, and return the components
-# indexed by i. Stolen from: 
-# http://stackoverflow.com/questions/15162741/what-is-rs-crossproduct-function
-crossprod_3d <- function(x, y, i = 1:3) {
+#' Compute the vector cross product between vectors x and y. Adapted from 
+#' http://stackoverflow.com/questions/15162741/what-is-rs-crossproduct-function
+#' @param x vector of length 3.
+#' @param y vector of length 3.
+#' @return vector cross product of x and y.
+#' @export
+crossprod_3d <- function(x, y) {
+  i <- 1:3
   # Project inputs into 3D, since the cross product only makes sense in 3D.
   To3D <- function(x) utils::head(c(x, rep(0, 3)), 3)
   x <- To3D(x)